Refactor: Centralize types in src/types directory
